New logic in new file
